pipeline/requirements.txt
h3>=4.0.0
geopandas>=0.14.0
shapely>=2.0.0
pandas>=2.0.0
pyarrow>=14.0.0
earthengine-api>=0.1.390
numpy>=1.24.0
google-cloud-storage>=2.14.0
rasterio>=1.3.0
tenacity>=8.2.0
